Enable restart, read_restart, and write_restart commands for MOLC
Quick fix to USER-MOLC/bond_ellipsoids.cpp to enable the use of
the restart command. The "not implemented" errors have been
removed. No new code as been addded. The bond style does not
actually write the bond coeffs to the restart file, or read them.
This is ok *IF the bond style and coeffs are fully specified in
the input file used for restarting the simulation*.